+++ src/VsPackage.cs
@@ -19,6 +19,21 @@ namespace MakeBddName
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
     [Guid(PackageGuids.guidPackageString)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
+    [ProvideOptionPage(
+        typeof(OptionsPage),
+        categoryName: "Make BDD Name",
+        pageName: "General",
+        categoryResourceID: 106,
+        pageNameResourceID: 120,
+        supportsAutomation: true)]
+    [ProvideProfile(
+        typeof(OptionsPage),
+        categoryName: "MakeBDDName",
+        objectName: "General",
+        categoryResourceID: 106,
+        objectNameResourceID: 107,
+        isToolsOptionPage: true,
+        DescriptionResourceID = 108)]
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1650:ElementDocumentationMustBeSpelledCorrectly",
          Justification = "pkgdef, VS and vsixmanifest are valid VS terms")]
     public sealed class VsPackage : Package
@@ -40,6 +55,12 @@ namespace MakeBddName
             return vsSelection != null ? new VsTextSelectionWrapper(vsSelection) : null;
         }
 
+        internal IOptions GetOptionsPage()
+        {
+            var page = (OptionsPage)GetDialogPage(typeof(OptionsPage));
+            return page;
+        }
+
         /// <summary>
         /// Initialization of the package; this method is called right after the package is sited, so
         /// this is the place where you can put all the initialization code that rely on services
@@ -54,7 +75,7 @@ namespace MakeBddName
             _dte = GetService(typeof(DTE)) as DTE2;
 
             // Initialize the commands.
-            MakeBddNameCommand.Initialize(menuCommandService, GetActiveSelection);
+            MakeBddNameCommand.Initialize(menuCommandService, GetActiveSelection, GetOptionsPage);
 
             Logger.LogDebug("Initialized");
         }